Fixed Visual Studio errors regarding basic numerical type definitions ( unsigned __int8 instead of __uint8 ); Fixed unrelated double cast warning on default linear and angular damping thresholds.

+++ ode/src/ode.cpp
@@ -1495,8 +1495,8 @@ dxWorld * dWorldCreate()
 
   w->dampingp.linear_scale = 0;
   w->dampingp.angular_scale = 0;
-  w->dampingp.linear_threshold = 0.01 * 0.01;
-  w->dampingp.angular_threshold = 0.01 * 0.01;  
+  w->dampingp.linear_threshold = REAL( 0.01 ) * REAL( 0.01 );
+  w->dampingp.angular_threshold = REAL( 0.01 ) * REAL( 0.01 );
   w->max_angular_speed = dInfinity;
 
   return w;
